Implemented dynamic ramp insertion
In `MCFR_HyS_BOP.ReactivityFeedback`, the variable `ReactExternalRamp` was redefigned so that the value was calculated from the parameters `ramp_rate` and `ramp_mag`. `ramp_rate` had its units changed to a new `ReactivityRate` type with units "1/s" and `ramp_mag` replaces the unused `ramp_low` and has units of "1".